Added a dummy CMLReader class... I.e., i did not implement any bindings yet, but the CML file is already parsed (and working :)
A test is given by org/openscience/cdk/test/CMLReaderTest.java. Invoke
it with all three jar files in the classpath.

Note that the CML import filter (cml.jar/JCFL) needs a so called
Chemical Data Object. This is the ChemFileCDO.java class and is
subclassed from ChemFile.java. So anything you like to do with a
ChemFile you can also do with ChemFileCDO.

Also i removed a misplaced call in CMLWriter.java.


git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@98 eb4e18e3-b210-0410-a6ab-dec725e4b171