Previously, unit tests in IAtomTypeTest would use code like:
IAtomType type = getBuilder().newAtomType()
// then, test e.g. getAtomTypeName()

This would test NNAtomType in the nonotify module, but it would only test the getAtomTypeName() method of NNAtomType.

Now, it is constructed with:

IAtomType type = (IAtomType)newChemObject();

which will actually return a NNAtomType, *or* subclass... and that's important, as that extending IAtomTypeTest
will now also test getAtomTypeName() of NNAtom, NNPseudoAtom, etc, which was not tested before.

git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/cdk/branches/cdk-1.2.x@13676 eb4e18e3-b210-0410-a6ab-dec725e4b171