It is not clear if the below patches are bug fixes, neither should @cdk.bug annotation be removed from unit tests.
Miguel, please use the patch system, as cdk1.2.x is in deep freeze, and all new functionality needs to be reviewed by
two reviewers (which should not really be a problem for the patches, but it just needs to happen)...

Bug fixing can generally go in directly and is reviewed by me via the commits mailing list.

Revert "Set value 1 when contains only the signal for specify the charge."
  This reverts commit 94613ba13effbf36d21c667bbcfc3f7b9701b183.
Revert "Correction for when the charge is positive. It is not possible to create a Integer from a string like "+1""
  This reverts commit 3cc4d3116dea04c67a483ad8a16632812bda18bd.
Revert "Capability to parser a molecular formula string containing charge signal. E.g. [O3S]2-"
  This reverts commit 46c4c9bf32c2236934685f4ef62b56695e7a9d51.
Revert "Removing @cdk.bug 2276507 from the javadoc. It is already closed"
  This reverts commit d2c3bb8e681ea4e0a3e91f1f3b0542e0a3ae4279.
Revert "Removing @cdk.bug tag 1944604 from the javadoc. It is already closed"
  This reverts commit df3ea843e0474057fd821c6fe7394ffadbb47f1c.
Revert "Removed cdk.bug tag fixed by Egonw"
  This reverts commit 44eff2d497f6b4582bb8148e81aff16f7b1f8a37.

git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/cdk/branches/cdk-1.2.x@14417 eb4e18e3-b210-0410-a6ab-dec725e4b171