Dont' configure the atom if it is a PseudoAtom
configuring aotm type information is not really valid for pseudo atoms - first because they basically have no type information and second because they may have information associated with them from another context, which should not be
overwritten. So we only do the stuff below if we have a non pseudoatom
        
a side effect of this is that it is probably not valid to get the atom type of a peudo atom. I think this is OK, since you can always check whether an atom is a pseudo atom without looking at its atom type

git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/cdk/branches/cdk-1.2.x@14438 eb4e18e3-b210-0410-a6ab-dec725e4b171