changed from no-op logger call to syserr
If we fail to read the cdk.debugging System property we can not know whether the user expects logging output or not. The old code seems to have the intent to inform the user about this, but it will never succeed with this cause the default value is to not have the logger print anything -- and since we failed to read the property we have not changed it. I think the reasonable action is to write a short error message about this to System.err and then nothing more.

Signed-off-by: Rajarshi  Guha <rajarshi.guha@gmail.com>