fix bug #3305550 - moved SMARTSQuery to where used
It was initialised staticly for the outer class. However since the SMARTSQueryTool has state it can not be shared among multiple FingerPrinters. Now every CountSubstructures instance will have it's own which should make it possible to run multiple PubchemFingerprinters in parallel.

Also, made the SMARTSQueryTool throw an IllegelargumentException instead of a checked CDKException if something is wrong with the SMARTS string.

Signed-off-by: Rajarshi  Guha <rajarshi.guha@gmail.com>