IteratingMDLReader Skip Patch
Provides skip functionality to IteratingMDLReader. If a molecule is null the reader
can skip over it (if 'skip' boolean is set true). The skip property is set via an
mutator 'setSkip' or in the constructor. If a null molecule is read the read will
skip forward until the next record separator ($$$$). The commit include a unit test
for the patch which uses some molecules from the HMDB sdf file.

Change-Id: Ie116eb1d5119357bec049a66e2d1a0a2a3dcb620
Signed-off-by: Egon Willighagen <egonw@users.sourceforge.net>