add the github site to the readme
Although I only added one sentence it might be nice to put the link the CDK github site in the CDK project description. For some reason, the github API docs are not the first google search when looking for online documentation so I feel this could be a little more prominent.