Correct logic for missing ligands in square planar (SP).
When we have missing ligands from SMARTS we need to insert placeholder (the central/focus atom) into the neighbour list. Since square planar has the same number as tetrahedral there was some incorrect logic about where the missing neighbours go. There was an optimisation for tetrahedrals with one missing neighbour but this logic was not correct for SP which also has 4 neighbours.