FIx #1183 InChITautomerGenerator on AtomContainer2
The InChI tautomer generator deleted atom/bonds but not stereochemistry leaving the 'skeleton' in an invalid state. The simple fix was to clear the stereochemistry from the skeleton (it was retained on the input molecule). Rather than delete the atoms one by one inefficiently it is better to just create the skeleton as we need it.